+++ .github/actions/compile-models/action.yml
@@ -38,7 +38,7 @@ runs:
       shell: bash
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install --upgrade transformers torch onnx onnxruntime onnxruntime-tools accelerate huggingface_hub pillow requests
+        python3 -m pip install -r ${{ github.action_path }}/requirements.txt
 
     - name: Prepare OpenVINO environment
       if: ${{ inputs.compiler-type == 'MLIR' }}
