+++ .github/actions/download-models/action.yml
@@ -22,7 +22,7 @@ runs:
       shell: bash
       run: |
         python3 -m pip install --upgrade pip
-        python3 -m pip install --upgrade huggingface_hub hf_xet requests
+        python3 -m pip install -r ${{ github.action_path }}/requirements.txt
 
     - name: Run downloader script
       shell: bash
