+++ .github/actions/download-models/requirements.txt
@@ -1,0 +1,3 @@
+huggingface_hub>=0.30.2
+hf_xet>=1.0.2
+requests>=2.32.3
